// src/types/models/mitsm/Marine.ts
export interface Marine {
  id: number;
  rank: string;
  first: string;
  last: string;
  edipi: string;
  mos: string;
  dor: string;
  afadbd: string;
  pebd: string;
  ecc: string;
  eas: string;
  dctb: string;
  rtd: string;
  fy: number;
  type_of_request: string;
  length_of_request: number;
  category: string;
  soe_code: string;
  organization: string;
  work_phone: string;
}